refactor(web): split shared/components/ui/VoiceMicButton.tsx into voice/
- VoiceMicButton.tsx (852 → 256 LOC) залишилось як публічний компонент;
  re-export `useVoiceInput`/`UseVoiceInputOptions`/`UseVoiceInputReturn`
  збережено для backward-compat (були публічні з 2025-Q4).
- voice/useVoiceInput.ts (139) — Web Speech API hook + types
  (`SpeechRecognitionLike`, `SpeechRecognitionCtor`,
  `getSpeechRecognitionCtor`).
- voice/useGroqVoiceInput.ts (270) — Groq Whisper recorder hook через
  `/api/transcribe` (старт/стоп MediaRecorder, upload, cleanup, hard-cap
  60s, мін-кап 250ms) + утиліти `pickRecorderMimeType`/`isGroqSupported`.
- voice/PendingVoiceChip.tsx (188) — 3-сек preview/undo чип з
  countdown-ring через requestAnimationFrame + Escape handler +
  viewport-positioning через portal у document.body.
- voice/resolveVoiceProvider.ts (12) — env-flag resolver
  (`VITE_VOICE_PROVIDER`: `auto`/`groq`/`webspeech`).

Поведінка не змінюється: ані конструктор MediaRecorder, ані список MIME
candidates (webm-opus → webm → mp4-aac → mp4 → ogg-opus → ogg), ані
текст помилок, ані 3-секундний таймер, ані позиціонування чипа,
ані deps useEffect/useCallback не зачіплено.

docs/tech-debt/frontend.md — оновлено count >600 LOC у apps/web/src
(22 → 21) + рекап декомпозиції.
